backend/requirements.txt
requests
sentence-transformers
scikit-learn
python-dotenv
numpy
fastapi
uvicorn
pydantic
diskcache